Improvement | Separation of utils functions to separate folder
