+++ index.html
@@ -20,6 +20,6 @@
       <button>Paper</button>
       <button>Scissors</button>
     </div>
-    <script src="script.js"></script>
+    <script src="script.js" type="module"></script>
   </body>
 </html>
