+++ modules/reverseMode.js
@@ -1,0 +1,6 @@
+const reverseMode = () => {
+  console.log(`Test`);
+  return document.querySelector("#score-container");
+};
+
+export { reverseMode };
